refactor(CreateAIReply): extract AI Response creation and updating to util function
